Move BDD tests into cukes directory
We're using the word `feature` in our domain so having a directory with the same name wasn't great.

I have renamed the usual Behave `features` directory to `cukes` to better reflect what's in that directory.
