+++ .behaverc
@@ -1,0 +1,2 @@
+[behave]
+paths=cukes
